# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gateway Documentation",
        "description": "This allows you to create,read, update and delete a product",
        "contact": {
            "email": "esekelvin24@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8181/api/v1",
            "description": "API Gateway Server"
        }
    ],
    "paths": {
...
        "/articles/{article_id}": {
            "get": {
                "tags": [
                    "View article by article ID"
                ],
                "summary": "View article using id",
                "description": "To display article when an article id is inserted",
                "operationId": "view_article",
                "parameters": [
                    {
                        "name": "article_id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "cover": {
                                            "type": "string",
                                            "example": "https://imgd.aeplcdn.com/1200x900/n/cw/ec/110233/2022-camry-exterior-right-front-three-quarter.jpeg?isig=0&q=75"
                                        },
                                        "full_text": {
                                            "type": "string",
                                            "example": "jdsndj  ksjdnkjn kjsnd"
                                        },
                                        "likes": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "views": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-01-22T13:33:14.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-01-22T13:33:14.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/articles/{article_id}/views": {
            "get": {
                "tags": [
                    "Show List of views"
                ],
                "summary": "Show List of Article views",
                "description": "Show list of views on an article",
                "operationId": "show_views",
                "parameters": [
                    {
                        "name": "article_id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Views",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "total number of views"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "View an article"
                ],
                "summary": "To view an article",
                "description": "View article",
                "operationId": "article_view",
                "parameters": [
                    {
                        "name": "article_id",
                        "in": "path",
                        "description": "Article id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "You just view this article",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "cover": {
                                            "type": "string",
                                            "example": "https://imgd.aeplcdn.com/1200x900/n/cw/ec/110233/2022-camry-exterior-right-front-three-quarter.jpeg?isig=0&q=75"
                                        },
                                        "full_text": {
                                            "type": "string",
                                            "example": "jdsndj  ksjdnkjn kjsnd"
                                        },
                                        "likes": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "views": {
                                            "type": "integer",
                                            "example": "0"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-01-22T13:33:14.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-01-22T13:33:14.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}